DAY-10 rental service

+++ src/main/java/com/kodilla/rental/domain/User.java
@@ -7,6 +7,7 @@ import lombok.*;
 import javax.persistence.*;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 
@@ -47,7 +48,7 @@ public class User {
     private String creditCardNo;
 
     @Column(name = "amount_to_pay")
-    private double toPay;
+    private BigDecimal toPay;
 
     @Column(name = "account_status")
     private boolean isBlocked;
